check for user to handle loading
