+++ src/components/HomeDashBoard.js
@@ -4,222 +4,10 @@ import React, { Component } from "react";
 import JournalDash from "./journal/JournalDash";
 import NotesDash from "./notes/NotesDash";
 
-const sampleUser = {
-  email: "johsn@smith.com",
-  name: "John Smith",
-  journal: {
-    weeks: [
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "this is my accomphlishment...",
-              learned: "that gurm is dumb...",
-              blockers: "gurm not pushing code",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "week 2 entry 1 accmom",
-              learned: "learned wk2 e1",
-              blockers: "blockers",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-      {
-        created: "2021-01-31T02:22:12+0000",
-        updated: "2021-01-31T02:22:12+0000",
-        entries: [
-          {
-            id: 1,
-            created: "2021-01-31T02:22:12+0000",
-            updated: "2021-01-31T02:22:12+0000",
-            content: {
-              accomplished: "",
-              learned: "",
-              blockers: "",
-            },
-          },
-        ],
-      },
-    ],
-  },
-  notes: [
-    {
-      header: "header",
-      content: "content",
-    },
-    {
-      header: "header",
-      content: "content",
-    },
-  ],
-};
-
 class HomeDashBoard extends Component {
-  constructor(props) {
-    super(props);
+  constructor() {
+    super();
     this.state = {
-      user: sampleUser,
       journalViewActive: true,
     };
 
@@ -231,7 +19,9 @@ class HomeDashBoard extends Component {
   }
 
   render() {
-    const { user, journalViewActive } = this.state;
+    const { journalViewActive } = this.state;
+    const { user } = this.props;
+
     return (
       <div>
         <div>
