+++ src/components/TextEditor.js
@@ -3,31 +3,28 @@ import { CKEditor } from "@ckeditor/ckeditor5-react";
 import ClassicEditor from "@ckeditor/ckeditor5-build-classic";
 
 class TextEditor extends React.Component {
-  constructor(props) {
-    super(props);
+  constructor() {
+    super();
     this.state = {};
   }
 
   render() {
-    const { defaultContents } = this.props;
-    let data = "<p>" + defaultContents + "</p>";
     return (
       <div onClick={this.focusEditor}>
         <CKEditor
           editor={ClassicEditor}
-          data={data}
+          data={"<p>Start your entry here...</p>"}
           onReady={(editor) => {
-            console.log("Editor is ready to use!", editor);
+            //console.log("Editor is ready to use!", editor);
           }}
           onChange={(event, editor) => {
-            const data = editor.getData();
-            console.log({ event, editor, data });
+            //console.log({ event, editor, data });
           }}
           onBlur={(event, editor) => {
-            console.log("Blur.", editor);
+            //console.log("Blur.", editor);
           }}
           onFocus={(event, editor) => {
-            console.log("Focus.", editor);
+            //console.log("Focus.", editor);
           }}
         />
       </div>
